requirements.txt
numpy==1.24.1
opencv_python_headless==4.7.0.68
Pillow==9.4.0
streamlit==1.17.0
tensorflow==2.11.0
tensorflow_intel==2.11.0